+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Change Log
 ==========
 
+2017-09-02
+----------
+
+ * re-added initialize/update timestamps to Timestampable
+
 2017-07-12
 ----------